# api-docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "RESTful API for managing census case data. Provides endpoints for querying and retrieving case information by various identifiers (UUID, reference, UPRN, postcode, QID). Returns comprehensive case details including address information, case status, and associated events.",
    "license" : {
      "name" : "Office for National Statistics",
      "url" : "https://www.ons.gov.uk"
    },
    "title" : "Case API Service",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost:8161",
    "description" : "Case API Service"
  } ],
...
  "paths" : {
...
    "/cases/qid/{qid}" : {
      "get" : {
        "description" : "Retrieves minimal case details linked to a specific questionnaire ID.",
        "operationId" : "findCaseByQid",
        "parameters" : [ {
          "description" : "Questionnaire Identifier",
          "in" : "path",
          "name" : "qid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaseContainerDTO"
                }
              }
            },
            "description" : "Case retrieved successfully"
          },
          "404" : {
            "description" : "QID not found"
          }
        },
        "summary" : "Find case by Questionnaire ID (QID)",
        "tags" : [ "Case Endpoint" ]
      }
    },
...
    "/qids/link" : {
      "put" : {
        "description" : "Links a questionnaire ID to a case. This endpoint is currently not implemented as the subscription for questionnaire links is not available.",
        "operationId" : "putQidLinkToCase",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewQidLink"
              }
            }
          },
          "description" : "QID link information to be created",
          "required" : true
        },
        "responses" : {
          "501" : {
            "description" : "Not Implemented - Questionnaire link subscription not available"
          }
        },
        "summary" : "Link QID to case",
        "tags" : [ "QID Endpoint" ]
      }
    },
    "/qids/{qid}" : {
      "get" : {
        "description" : "Retrieves the UAC-QID link information for a specified questionnaire ID, including associated case ID if available.",
        "operationId" : "getUacQidLinkByQid",
        "parameters" : [ {
          "description" : "Questionnaire Identifier (QID) to retrieve",
          "in" : "path",
          "name" : "qid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QidLink"
                }
              }
            },
            "description" : "QID link details retrieved successfully"
          },
          "404" : {
            "description" : "Questionnaire ID not found"
          },
          "500" : {
            "description" : "Internal server error occurred"
          }
        },
        "summary" : "Retrieve QID link details",
        "tags" : [ "QID Endpoint" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "CaseContainerDTO" : {
        "required" : [ "caseRef", "id" ],
        "type" : "object",
        "properties" : {
          "abpCode" : {
            "type" : "string",
            "description" : "AddressBase classification code",
            "example" : "RD06"
          },
          "addressLevel" : {
            "type" : "string",
            "description" : "Address level classification (e.g., E, U (signifying Establishment and Unit))",
            "example" : "U"
          },
          "addressLine1" : {
            "type" : "string",
            "description" : "First address line",
            "example" : "Flat 51 Francombe House"
          },
          "addressLine2" : {
            "type" : "string",
            "description" : "Second address line",
            "example" : "Commercial Road"
          },
          "addressLine3" : {
            "type" : "string",
            "description" : "Third address line"
          },
          "addressType" : {
            "type" : "string",
            "description" : "Residential address frame type (e.g., HH, CE)",
            "example" : "HH"
          },
          "caseEvents" : {
            "type" : "array",
            "description" : "List of events associated with the case",
            "items" : {
              "$ref" : "#/components/schemas/CaseEventDTO"
            }
          },
          "caseRef" : {
            "type" : "string",
            "description" : "Unique numeric reference for the case",
            "example" : "100000000000001"
          },
          "caseType" : {
            "type" : "string",
            "description" : "Case type classification (e.g., HH, HI, CE). It will match addressType unless it is an individual (HI) case.",
            "example" : "HH"
          },
          "collectionExerciseId" : {
            "type" : "string",
            "description" : "Collection Exercise UUID identifier",
            "format" : "uuid",
            "example" : "b22e3456-e89b-12d3-a456-426614174000"
          },
          "createdDateTime" : {
            "type" : "string",
            "description" : "Date and time when the case was created",
            "format" : "date-time",
            "example" : "2024-01-15T10:30:00Z"
          },
          "estabType" : {
            "type" : "string",
            "description" : "Establishment type (e.g., HALL OF RESIDENCE, HOUSEHOLD, SHELTERED ACCOMMODATION, RESIDENTIAL CARAVAN, RESIDENTIAL BOAT)",
            "example" : "HOUSEHOLD"
          },
          "estabUprn" : {
            "type" : "string",
            "description" : "Establishment UPRN for non-household establishments",
            "example" : "10008677190"
          },
          "id" : {
            "type" : "string",
            "description" : "Unique case UUID identifier",
            "format" : "uuid",
            "example" : "a11e3456-e89b-12d3-a456-426614174000"
          },
          "invalid" : {
            "type" : "boolean",
            "description" : "Flag indicating if the case record is marked as invalid",
            "example" : false
          },
          "lad" : {
            "type" : "string",
            "description" : "Local Authority District code (e.g., N06000023, S06000023, E06000023, W06000023)",
            "example" : "E06000023"
          },
          "lastUpdated" : {
            "type" : "string",
            "description" : "Date and time when the case was last updated",
            "format" : "date-time",
            "example" : "2024-01-20T14:45:00Z"
          },
          "latitude" : {
            "type" : "string",
            "description" : "Geographic latitude coordinate",
            "example" : "51.5074"
          },
          "longitude" : {
            "type" : "string",
            "description" : "Geographic longitude coordinate",
            "example" : "-0.1278"
          },
          "lsoa" : {
            "type" : "string",
            "description" : "Lower Layer Super Output Area grid reference (e.g., N01014540, S01014540, E01014540, W01014540)",
            "example" : "E01014540"
          },
          "msoa" : {
            "type" : "string",
            "description" : "Middle Layer Super Output Area grid reference (e.g., N02003043, S02003043, E02003043, W02003043)",
            "example" : "E02003043"
          },
          "oa" : {
            "type" : "string",
            "description" : "Output Area grid reference (e.g., N00073438, S00073438, E00073438, W00073438)",
            "example" : "E00073438"
          },
          "organisationName" : {
            "type" : "string",
            "description" : "Name of the organisation at the address",
            "example" : "Acme Corporation"
          },
          "postcode" : {
            "type" : "string",
            "description" : "UK postal code",
            "example" : "XX1 0XX"
          },
          "region" : {
            "type" : "string",
            "description" : "Administrative region code (e.g., N12000007, S12000007, E12000007, W12000007)",
            "example" : "E12000007"
          },
          "secureEstablishment" : {
            "type" : "boolean",
            "description" : "Indicator whether address is a secure establishment",
            "example" : false
          },
          "surveyType" : {
            "type" : "string",
            "description" : "Type of survey (e.g., CENSUS, CCS)",
            "example" : "CENSUS"
          },
          "townName" : {
            "type" : "string",
            "description" : "Town or city name",
            "example" : "Windleybury"
          },
          "uprn" : {
            "type" : "string",
            "description" : "Unique Property Reference Number",
            "example" : "10008677190"
          }
        },
        "description" : "Data Transfer Object representing a census case container",
        "example" : {
          "caseRef" : "100000000000001",
          "id" : "a11e3456-e89b-12d3-a456-426614174000",
          "estabType" : "HOUSEHOLD",
          "uprn" : "10008677190",
          "estabUprn" : "10008677190",
          "collectionExerciseId" : "b22e3456-e89b-12d3-a456-426614174000",
          "surveyType" : "CENSUS",
          "addressType" : "HH",
          "caseType" : "HH",
          "createdDateTime" : "2024-01-15T10:30:00Z",
          "addressLine1" : "Flat 51 Francombe House",
          "addressLine2" : "Commercial Road",
          "addressLine3" : "Suite 3",
          "townName" : "Windleybury",
          "postcode" : "XX1 0XX",
          "organisationName" : "Acme Corporation",
          "addressLevel" : "U",
          "abpCode" : "RD06",
          "region" : "E12000007",
          "latitude" : "51.5074",
          "longitude" : "-0.1278",
          "oa" : "E00073438",
          "lsoa" : "E01014540",
          "msoa" : "E02003043",
          "lad" : "E06000023",
          "lastUpdated" : "2024-01-20T14:45:00Z",
          "caseEvents" : [ ],
          "secureEstablishment" : false,
          "invalid" : false
        }
      },
...
      "CaseEventDTO" : {
        "required" : [ "createdDateTime", "eventType", "id" ],
        "type" : "object",
        "properties" : {
          "createdDateTime" : {
            "type" : "string",
            "description" : "Date and time when the event was created",
            "format" : "date-time",
            "example" : "2024-01-15T10:30:00Z"
          },
          "description" : {
            "type" : "string",
            "description" : "Human-readable description of the event",
            "example" : "Case created for collection exercise"
          },
          "eventType" : {
            "$ref" : "#/components/schemas/EventTypeDTO"
          },
          "id" : {
            "type" : "string",
            "description" : "Unique event identifier (UUID)",
            "format" : "uuid",
            "example" : "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          }
        },
        "description" : "Data Transfer Object representing an event associated with a case",
        "example" : {
          "id" : "f47ac10b-58cc-4372-a567-0e02b2c3d479",
          "eventType" : "NEW_CASE",
          "description" : "Case created for collection exercise",
          "createdDateTime" : "2024-01-15T10:30:00Z"
        }
      },
      "EventTypeDTO" : {
        "type" : "string",
        "description" : "Enumeration of all possible event types that can occur in the census RM system",
        "example" : "NEW_CASE",
        "enum" : [ "NEW_CASE", "RECEIPT", "REFUSAL", "INVALID_CASE", "EQ_LAUNCH", "UAC_AUTHENTICATION", "PRINT_FULFILMENT", "EXPORT_FILE", "DEACTIVATE_UAC", "UPDATE_SAMPLE", "UPDATE_SAMPLE_SENSITIVE", "SMS_FULFILMENT", "ACTION_RULE_SMS_REQUEST", "EMAIL_FULFILMENT", "ACTION_RULE_EMAIL_REQUEST", "ACTION_RULE_SMS_CONFIRMATION", "ACTION_RULE_EMAIL_CONFIRMATION", "ERASE_DATA" ]
      },
      "NewQidLink" : {
        "type" : "object",
        "properties" : {
          "channel" : {
            "type" : "string",
            "description" : "Channel through which this link was created (e.g., CONTACT_CENTRE, PHONE, ONLINE)",
            "example" : "CONTACT_CENTRE"
          },
          "qidLink" : {
            "$ref" : "#/components/schemas/QidLink"
          },
          "transactionId" : {
            "type" : "string",
            "description" : "Transaction ID for tracing this QID link request",
            "format" : "uuid",
            "example" : "a11e3456-e89b-12d3-a456-426614174000"
          }
        },
        "description" : "Data Transfer Object for creating a new QID link to a case",
        "example" : {
          "transactionId" : "a11e3456-e89b-12d3-a456-426614174000",
          "channel" : "CONTACT_CENTRE",
          "qidLink" : {
            "questionnaireId" : "Q123456",
            "caseId" : "a11e3456-e89b-12d3-a456-426614174000"
          }
        }
      },
      "QidLink" : {
        "required" : [ "questionnaireId" ],
        "type" : "object",
        "properties" : {
          "caseId" : {
            "type" : "string",
            "description" : "UUID of the census case associated with this QID",
            "format" : "uuid",
            "example" : "a11e3456-e89b-12d3-a456-426614174000"
          },
          "questionnaireId" : {
            "type" : "string",
            "description" : "Unique Questionnaire Identifier (QID) assigned to the respondent",
            "example" : "Q123456"
          }
        },
        "description" : "Data Transfer Object representing the link between a Questionnaire ID (QID) and a Case",
        "example" : {
          "questionnaireId" : "Q123456",
          "caseId" : "a11e3456-e89b-12d3-a456-426614174000"
        }
      }
    }
  }
}